validator/requirements.txt
pandas==0.24.1
dnspython==1.16.0
